% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brd_mat_2021.R
\name{brd.mat}
\alias{brd.mat}
\title{Generate a Rule-Based Breeding Matrix}
\usage{
brd.mat(moms = 10, dads = 10, lambda.low = 1, lambda.high = 1)
}
\arguments{
\item{moms}{The number of female parents in the population.}

\item{dads}{The number of male parents in the population.}

\item{lambda.low}{The lower bound of the lambda value for the Poisson
distribution used to determine the number of mates.}

\item{lambda.high}{The upper bound of the lambda value for the Poisson
distribution used to determine the number of mates.}
}
\value{
A binary breeding matrix with \code{dads} rows and \code{moms} columns, where
a '1' indicates a mating occurred.
}
\description{
This function simulates a breeding matrix based on a set of
rules and a Poisson distribution for the number of mates per individual.
This method is known to introduce a systematic bias due to the order of
operations and is used here to demonstrate the superiority of a more
rigorous approach.
}
\examples{
# Generate a simple breeding matrix
my_matrix <- brd.mat(moms = 50, dads = 50, lambda.low = 1, lambda.high = 1)
# Note: This is for demonstration of the old method's internal biases.
}
