% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_target_viability_helper.R
\name{check_target_viability}
\alias{check_target_viability}
\title{Check Demographic Target Viability}
\usage{
check_target_viability(
  sr_target,
  mean_mates_target,
  max_males_per_female,
  max_females_per_male
)
}
\arguments{
\item{sr_target}{The target operational sex ratio (Males / Females).}

\item{mean_mates_target}{The target global mean mates.}

\item{max_males_per_female}{Biological cap on how many males a female can mate with.}

\item{max_females_per_male}{Biological cap on how many females a male can mate with.}
}
\value{
TRUE if viable. Stops execution with an informative error if impossible.
}
\description{
Evaluates whether a requested combination of Sex Ratio and Mean Mates
is mathematically possible given the biological mating caps of the species.
}
