% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mateR2-package.R
\docType{package}
\name{mateR2-package}
\alias{mateR2}
\alias{mateR2-package}
\title{mateR2: Simulate Single-Cohort Pedigrees Using a Constrained Bipartite Sampler}
\description{
Simulates single-cohort pedigrees for dioecious species by treating the mating pool as a bipartite network. A Metropolis-Hastings Markov chain Monte Carlo sampler, written in C++ via 'Rcpp', navigates a compressed state space of discrete mating-block configurations to satisfy user-specified targets for parental abundance, operational sex ratio and mean number of mates. A joint-degree-preserving edge-swapping routine then rewires the expanded individual-level matrix to dissolve modular block boundaries while holding every individual's mating success constant. Fecundity allocation, offspring sub-sampling, sibship and parental yield-class statistics support benchmarking of pedigree reconstruction software and evaluation of field sampling designs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicksard/mateR2}
  \item Report bugs at \url{https://github.com/nicksard/mateR2/issues}
}

}
\author{
\strong{Maintainer}: Nicholas M. Sard \email{nicksard@gmail.com} (\href{https://orcid.org/0000-0001-7695-2946}{ORCID})

Authors:
\itemize{
  \item Nicholas M. Sard \email{nicksard@gmail.com} (\href{https://orcid.org/0000-0001-7695-2946}{ORCID})
}

}
\keyword{internal}
