% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior_ensemble.R
\name{sample_posterior_ensemble}
\alias{sample_posterior_ensemble}
\title{Extract High-Fidelity Posterior Ensemble of Mate-Pair Count Vectors}
\usage{
sample_posterior_ensemble(
  mcmc_res,
  target_Np,
  target_SR,
  target_MM,
  burn_in,
  thin,
  n_samples = 100,
  max_error_pct = 0.05
)
}
\arguments{
\item{mcmc_res}{List object returned by generate_map_table().}

\item{target_Np}{Numeric. Your target for total parents.}

\item{target_SR}{Numeric. Your target sex ratio.}

\item{target_MM}{Numeric. Your target mean mates.}

\item{burn_in}{Integer. The burn-in used during the MCMC run.}

\item{thin}{Integer. The thinning interval used during the MCMC run.}

\item{n_samples}{Integer. Number of posterior draws to retain (default: 100).}

\item{max_error_pct}{Numeric. Max allowable relative Euclidean error (default: 0.05).}
}
\description{
Extract High-Fidelity Posterior Ensemble of Mate-Pair Count Vectors
}
