% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_map_table.R
\name{generate_map_table}
\alias{generate_map_table}
\title{Run and Analyze the MCMC Breeding Matrix Sampler}
\usage{
generate_map_table(
  Np_target,
  sr_target,
  mean_mates_target,
  max_males_per_female,
  max_females_per_male,
  decay_constant = -0.5,
  np_weight = 10,
  sr_weight = 1,
  mm_weight = 10,
  n_iter = 2e+05,
  burn_in = 20000,
  thin = 20,
  initial_method = "auto",
  seed = NULL
)
}
\arguments{
\item{Np_target}{The target total number of parents (males + females).}

\item{sr_target}{The target sex ratio (males / females).}

\item{mean_mates_target}{The target mean number of mates per individual.}

\item{max_males_per_female}{The maximum number of males a female can mate with.}

\item{max_females_per_male}{The maximum number of females a male can mate with.}

\item{decay_constant}{A negative value that penalizes complex mating structures.
A larger negative value results in a higher penalty.}

\item{np_weight}{The weight given to the target Number of Parents (Np).}

\item{sr_weight}{The weight given to the target Sex Ratio (SR).}

\item{mm_weight}{The weight given to the target Mean Mates (MM).}

\item{n_iter}{The total number of iterations for the MCMC sampler.}

\item{burn_in}{The number of initial iterations to discard before sampling begins.}

\item{thin}{The thinning interval for collecting samples from the MCMC chain.}

\item{initial_method}{A character string ("auto") to dynamically calculate the
optimal starting state based on targets, or a numeric vector representing a
custom starting state of block counts.}

\item{seed}{An optional seed for the random number generator to ensure
reproducibility.}
}
\value{
A list containing the MAP estimate table, the summary statistics#'   for the MAP estimate, and the raw MCMC output.
}
\description{
This is the main wrapper function that orchestrates the entire MCMC
simulation. It sets up the initial conditions, runs the core C++ sampler,
analyzes the output to find the Maximum a Posteriori (MAP) estimate, and
calculates key statistics for the resulting breeding matrix.
}
