% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mateR2-package.R
\docType{package}
\name{mateR2-package}
\alias{mateR2}
\alias{mateR2-package}
\title{mateR2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
