% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_table_to_matrix.R
\name{mp_table_to_matrix}
\alias{mp_table_to_matrix}
\title{Convert MCMC Mate Pair Table to a Breeding Matrix}
\usage{
mp_table_to_matrix(mp_table)
}
\arguments{
\item{mp_table}{A data frame with at least the columns \code{Males}, \code{Females}, and \code{Count},
representing the number of males, females, and the number of blocks of that
type.}
}
\value{
A binary matrix where rows represent males and columns represent females.
A value of 1 indicates a mating occurred. The matrix is block-diagonal.
}
\description{
This function takes a summary table of breeding configurations
(e.g., the output from a MCMC run) and converts it into a structured, binary
Individual-by-Individual Matrix (IIM).
}
