% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped2mat.R
\name{ped2mat}
\alias{ped2mat}
\title{Convert a Pedigree Data Frame to a Breeding Matrix}
\usage{
ped2mat(ped)
}
\arguments{
\item{ped}{A data frame representing a pedigree, with columns \code{dad} and \code{mom}.}
}
\value{
A matrix where the cell at \verb{[i, j]} contains the number of offspring
produced by male \code{i} and female \code{j}.
}
\description{
This function takes a pedigree in a long data frame format and
converts it into a wide breeding matrix (Individual-by-Individual Matrix),
where rows represent males (dads) and columns represent females (moms).
}
